Update Schematic
Here the design for the PCB is more or less in a good spot. The gate drivers may get expanded some to take up more room (to make it easier to tell which components go where in the PCB). Most of the traces are placed, though this may change in the future.