First PCB Final
Here the PCB is fully connected. I do not like the design, but it is fully connected either way. The PCB is still missing the current sensor, as I have no way of knowing what components the small boards we have off amazon are using.

If I had more time, this design would definitely be reworked.